% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pigeon_importr.R
\name{pigeon_import}
\alias{pigeon_import}
\title{Batch imports Datavyu, Habit, or Director Data}
\usage{
pigeon_import(method = "default", pattern.regex = NULL,
  pattern.exc = NULL, path = getwd())
}
\arguments{
\item{method}{Type of raw data being clean}

\item{pattern.regex}{Regular expression being searched}

\item{pattern.exc}{Regular expression being excluded}

\item{path}{File location}
}
\value{
Raw data of a single type compiled in a large list with each element being a unique participant
}
\description{
Batch imports Datavyu, Habit, or Director Data
}
\examples{
pigeon_import("datavyu", "NumbR10", "Number Replication")
pigeon_import("habit", "Number Replication")
pigeon_import("director", "MR10", ".txt", "./data")
}
