% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pigeon_processr.R
\name{pigeon_process}
\alias{pigeon_process}
\title{Processes and combines cleaned data}
\usage{
pigeon_process(x, method = "default", exports = TRUE, endformat = "wide",
  join = "full")
}
\arguments{
\item{x}{A list of data to be processed and cleaned}

\item{method}{A vector listing types of input data}

\item{exports}{Logical or 1:3, determines what csv.s are exported and when}

\item{endformat}{"wide" or "long"}

\item{join}{"full","left","right","" determines how data will be combined}
}
\value{
A dataframe with aggregated looking, widened data, and trial information
}
\description{
Processes and combines cleaned data
}
