% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{pigeontools_deps}
\alias{pigeontools_deps}
\title{List all pigeontools dependencies}
\usage{
pigeontools_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
pigeontools packages.}
}
\description{
List all pigeontools dependencies
}
