% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{pigeontools_packages}
\alias{pigeontools_packages}
\title{List all packages in the pigeontools}
\usage{
pigeontools_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include pigeontools in the list?}
}
\description{
List all packages in the pigeontools
}
\examples{
pigeontools_packages()
}
